requirements.txt
torch>=2.1
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
streamlit>=1.34
folium>=0.15
streamlit-folium>=0.18
torchvision>=0.16
Pillow>=10.0
fastapi>=0.110
uvicorn>=0.24
geopandas>=0.14
branca>=0.7
